requirements.txt
numpy
opencv-python
opencv-contrib-python
pyrealsense2
scikit-learn
PyYAML